feat: rename rco3lib to librco3 + actions stuff
