+++ .github/workflows/build.yml
@@ -87,6 +87,39 @@ jobs:
         with:
           name: rco3
           path: RCO.zip
+  build-librco3:
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v3
+      - name: Install pnpm
+        uses: pnpm/action-setup@v2
+        with:
+          version: 'latest'
+      - name: Install Node.js
+        uses: actions/setup-node@v3
+        with:
+          node-version: 20
+          cache: 'pnpm'
+          cache-dependency-path: 'LibRCO3/pnpm-lock.yaml'
+      - name: Install dependencies
+        run: |
+          cd LibRCO3
+          pnpm install
+      - name: Build Site
+        run: |
+          cd LibRCO3
+          pnpm build
+      - name: Upload RCO3 Library
+        uses: actions/upload-artifact@v3
+        with:
+          name: librco3-dist
+          path: LibRCO3/dist
+      - name: Upload RCO3 Docs
+        uses: actions/upload-artifact@v3
+        with:
+          name: librco3-docs
+          path: LibRCO3/docs
   upload-static:
     runs-on: ubuntu-latest
     steps:
@@ -104,7 +137,7 @@ jobs:
           path: assets
   merge-builds:
     runs-on: ubuntu-latest
-    needs: [build-rco3-installer, build-rco3-site, build-rco3, upload-static]
+    needs: [build-rco3-installer, build-rco3-site, build-rco3, build-librco3, upload-static]
     steps:
       - name: Setup Pages
         uses: actions/configure-pages@v3
@@ -123,6 +156,16 @@ jobs:
         with:
           name: rco3
           path: .
+      - name: Download LibRCO3
+        uses: actions/download-artifact@v3
+        with:
+          name: librco3-dist
+          path: librco3-dist
+      - name: Download LibRCO3 Documentation
+        uses: actions/download-artifact@v3
+        with:
+          name: librco3-docs
+          path: librco3-docs
       - name: Unzip
         run: |
           unzip -u RCO.zip
