+++ LibRCO3/package.json
@@ -58,6 +58,6 @@
   "repository": {
     "type": "git",
     "url": "https://github.com/L8X/Roblox-Client-Optimizer.git",
-    "directory": "RCO3Lib"
+    "directory": "LibRCO3"
   }
 }
